+++ src/http/adapter.js
@@ -22,16 +22,21 @@ function buildActionUrls( state, resourceName, actionName, action, resource, res
 	} else {
 		source = action.url || [ "" ];
 	}
-	return _.map( source, function( actionUrl ) {
+	var results = _.flatten( _.map( source, function( actionUrl ) {
 		if ( _.isRegExp( actionUrl ) ) {
 			return regex.prefix(
 				state.http.buildUrl( state.config.urlPrefix || '', prefix ),
 				actionUrl
 			);
 		} else if ( state.config.urlStrategy ) {
-			var url = state.config.urlStrategy( resourceName, actionName, action, resources );
-			prefix = hasPrefix( state, url ) ? '' : prefix;
-			return state.http.buildUrl( prefix, url );
+			var urls = state.config.urlStrategy( resourceName, actionName, action, resources );
+			if( _.isString( urls ) ) {
+				urls = [ urls ];
+			}
+			return _.map( urls, function( url ) {
+				prefix = hasPrefix( state, url ) ? '' : prefix;
+				return state.http.buildUrl( prefix, url );
+			} );
 		} else {
 			var resourceIndex = actionUrl ? actionUrl.indexOf( resourceName ) : -1;
 			var resourcePrefix = resourceIndex === 0 || resourceIndex === 1 ? '' : resourceName;
@@ -39,7 +44,8 @@ function buildActionUrls( state, resourceName, actionName, action, resource, res
 				prefix, resource.urlPrefix || '', resourcePrefix, ( actionUrl || '' )
 			);
 		}
-	} );
+	} ) );
+	return results;
 }
 
 function buildActionAlias( resourceName, actionName ) {
@@ -179,12 +185,54 @@ function getHandler( handle ) {
 	}
 }
 
+function getAuthorize( resource, action, envelope ) {
+	var authorize = action.authorize;
+	if( _.isFunction( authorize ) ) {
+		return authorize;
+	} else {
+		var list = _.filter( _.map( authorize, function( option ) {
+			if( _.isFunction( option.when ) ) {
+				return option;
+			} else if( option.when === true ) {
+				return {
+					when: function() { return true; },
+					then: option.then
+				};
+			} else if( _.isObject( option.when ) ) {
+				return {
+					when: _.matches( option.when ),
+					then: option.then
+				};
+			} else {
+				console.error(
+					format(
+						'Differentiated authorize\'s \'when\' property must be a function or an object instead of \'%s\'. Option will not be included in potential outcomes.'
+					), 
+					option.when
+				);
+			}
+		} ) );
+		return function( envelope ) {
+			var option = _.find( list, function( option ) {
+				return option.when( envelope );
+			} );
+			if( option ) {
+				return option.then( envelope );	
+			} else {
+				return { status: 403, data: { message: 'User lacks sufficient permissions' } };
+			}
+		};
+	}
+}
+
 function getPermissionCheck( state, meta, resource, action, envelope ) {
 	var check;
 	if( action.authorize ) {
 		check = function() {
-			var can = action.authorize.bind( resource )( envelope, envelope.context );
+			var authCall = getAuthorize( resource, action, envelope );
+			var can = authCall.bind( resource )( envelope, envelope.context );
 			return can && can.then ? can : when.resolve( can );
+
 		};
 		return checkPermissionFor.bind( undefined, state, check, meta, action );
 	} else if( state.auth && state.auth.checkPermission ) {
@@ -294,9 +342,7 @@ function wireupResource( state, resource, basePath, resources ) {
 function wireupAction( state, resource, actionName, action, metadata, resources ) {
 	var meta = getActionMetadata( state, resource, actionName, action, metadata, resources );
 	log.debug( 'Mapping resource \'%s\' action \'%s\' to %s %s',
-		resource.name, actionName, action.method, meta.url );
-
-	console.log( "	META", meta.urls );
+		resource.name, actionName, action.method, meta.urls.join( ", " ) );
 	_.each( meta.urls, function( url ) {
 		state.http.route( url, action.method, function( req, res ) {
 			meta.getEnvelope( req, res );